+++ core/services/chainlink/application_test.go
@@ -7,6 +7,8 @@ import (
 	"testing"
 
 	"github.com/smartcontractkit/chainlink/core/internal/cltest"
+	"github.com/smartcontractkit/chainlink/core/services/chainlink"
+	"github.com/stretchr/testify/assert"
 
 	"github.com/onsi/gomega"
 	"github.com/stretchr/testify/require"
@@ -31,4 +33,18 @@ func TestChainlinkApplication_SignalShutdown(t *testing.T) {
 	gomega.NewGomegaWithT(t).Eventually(func() bool {
 		return completed.IsSet()
 	}).Should(gomega.BeTrue())
+}
+
+func TestChainlinkApplication_ChangeInChainID(t *testing.T) {
+	config := cltest.NewTestEVMConfig(t)
+	app, cleanup := cltest.NewApplicationWithConfig(t, config)
+	defer cleanup()
+
+	require.NoError(t, app.Store.ORM.DB.Exec(`
+INSERT INTO configurations (name, value, updated_at, created_at)
+VALUES('ETH_CHAIN_ID', '2663', now(), now());`,
+	).Error)
+	config.GeneralConfig.Overrides.SetChainID(7853)
+	err := app.Start()
+	assert.Equal(t, chainlink.ErrNewChainID, err)
 }