+++ core/web/log_controller_test.go
@@ -37,7 +37,7 @@ func TestLogController_GetLogConfig(t *testing.T) {
 
 	cfg := cltest.NewTestEVMConfig(t)
 	cfg.GeneralConfig.Overrides.EthereumDisabled = null.BoolFrom(true)
-	logLevel := config.LogLevel{zapcore.WarnLevel}
+	logLevel := config.LogLevel{Level: zapcore.WarnLevel}
 	cfg.GeneralConfig.Overrides.LogLevel = &logLevel
 	sqlEnabled := true
 	cfg.GeneralConfig.Overrides.LogSQLStatements = null.BoolFrom(sqlEnabled)